Don't require root path in assisted chat metrics
Currently the overview dashboard for the assisted chat include a label
constraint like `path="/"` in the Prometheus queries that it uses. But
those metrics aren't cumulative, they are path specific, and they will
always be zero for the root path. As a result the dashboard is always
empty. This patch changes the dashboard so that it will show the paths
aggregated by default, but also give the user the possibility to click
in a specific path, in particular in the `/v1/query` and
`/v1/streaming_query` paths, which are the ones releveant for the user.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>